% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.stack.R
\name{extend.stack}
\alias{extend.stack}
\title{Extend your jenga stack}
\usage{
extend.stack(stack, extendedSlot)
}
\arguments{
\item{stack}{A jenga stack as created by \code{create.slot}}

\item{extendedSlot}{a character input, of the name of the new slot to be created.}
}
\value{
an extended jenja stack
}
\description{
Extend your jenga stack
}
