% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_stack.R
\name{extend_stack}
\alias{extend_stack}
\title{Extend your jenga stack}
\usage{
extend_stack(stack, new.slot)
}
\arguments{
\item{stack}{A jenga stack as created by \code{create.slot}}

\item{new.slot}{a character input, of the name of the new slot to be created.}
}
\value{
an extended jenja stack
}
\description{
Extend your jenga stack
}
