% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hood.R
\name{hood}
\alias{hood}
\title{support functions for jenga}
\usage{
hood(
  ts,
  seq_len,
  k,
  method,
  kernel,
  ci,
  deriv,
  n_windows,
  mode,
  dates,
  error_scale,
  error_benchmark
)
}
\arguments{
\item{ts}{A data frame.}

\item{seq_len}{Positive integer.}

\item{k}{Positive integer.}

\item{method}{Positive integer.}

\item{kernel}{String.}

\item{ci}{Confidence interval.}

\item{deriv}{Integer vector.}

\item{n_windows}{Positive integer.}

\item{mode}{String.}

\item{dates}{Date.}

\item{error_scale}{String.}

\item{error_benchmark}{String.}
}
\description{
support functions for jenga
}
\author{
Giancarlo Vercellino \email{giancarlo.vercellino@gmail.com}
}
