% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.stack.R
\name{merge.stack}
\alias{merge.stack}
\title{Merge your jenga stacks}
\usage{
merge.stack(stack1, stack2)
}
\arguments{
\item{stack1}{A jenga stack as created by \code{create.slot}}

\item{stack2}{A jenga stack as created by \code{create.slot}}
}
\value{
a merged jenja stack
}
\description{
Merge your jenga stacks
}
