% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.jenga_plot}
\alias{plot.jenga_plot}
\title{Plot jenga forecast objects}
\usage{
\method{plot}{jenga_plot}(x, ...)
}
\arguments{
\item{x}{A jenga forecast plot object.}

\item{...}{Additional arguments passed to the base plotting device.}
}
\description{
Plot jenga forecast objects
}
