% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.jenga_plot}
\alias{print.jenga_plot}
\title{Print jenga forecast plot objects}
\usage{
\method{print}{jenga_plot}(x, ...)
}
\arguments{
\item{x}{A jenga forecast plot object.}

\item{...}{Additional arguments passed to the base plotting device.}
}
\description{
Print jenga forecast plot objects
}
