% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_stack.R
\name{status_stack}
\alias{status_stack}
\title{Check the status of your jenga stack}
\usage{
status_stack(stack)
}
\arguments{
\item{stack}{A jenga stack as created by \code{create.slot}}
}
\value{
print statement on the number of filled and empty slots in your jenga stack
}
\description{
Check the status of your jenga stack
}
