% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.stack.R
\name{update.stack}
\alias{update.stack}
\title{Update a slot of your jenga stack}
\usage{
update.stack(stack, objects, slots, save = TRUE)
}
\arguments{
\item{stack}{A jenga stack as created by \code{create.slot}}

\item{objects}{a list of items to be inserted into the jenga stack}

\item{slots}{a vector of character inputs, of the slot/s to be updated. Must match slots already in the stack. Order of slots corresponds to order of objects}

\item{SAVE}{if \code{SAVE=TRUE}, then the updated slot will be saved to disk in an .RData file of the name of the stack.}
}
\value{
an updated jenja stack
}
\description{
Update a slot of your jenga stack
}
