% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{engine}
\alias{engine}
\title{support functions for jenga}
\usage{
engine(
  ts,
  seq_len,
  k,
  method,
  kernel,
  error_measurement,
  deriv,
  mode,
  error_scale,
  error_benchmark
)
}
\arguments{
\item{ts}{A data frame.}

\item{seq_len}{Positive integer.}

\item{k}{Positive integer.}

\item{method}{Positive integer.}

\item{kernel}{String.}

\item{error_measurement}{Logical.}

\item{deriv}{Integer vector.}

\item{mode}{String.}

\item{error_scale}{String.}

\item{error_benchmark}{String.}
}
\description{
support functions for jenga
}
\author{
Giancarlo Vercellino \email{giancarlo.vercellino@gmail.com}
}
