% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.stack.R
\name{update.stack}
\alias{update.stack}
\title{Update a slot of your jenga stack}
\usage{
update.stack(stack, object, slot, save = TRUE)
}
\arguments{
\item{stack}{A jenga stack as created by \code{create.slot}}

\item{object}{An object to be inserted into the jenga stack}

\item{updatedSlot}{a character input, of the slot to be updated. Must match slots already in the stack.}

\item{SAVE}{if \code{SAVE=TRUE}, then the updated slot will be saved to disk in an .RData file of the name of the stack.}
}
\value{
an updated jenja stack
}
\description{
Update a slot of your jenga stack
}
